requirements.txt
datasets==3.3.2
faiss_cpu==1.10.0
Flask==3.1.0
flask_cors==5.0.1
langchain==0.3.25
langchain_ollama==0.3.2
matplotlib==3.10.1
nltk==3.9.1
numpy==2.2.5
rank_bm25==0.2.2
scikit_learn==1.6.1
sentence_transformers==3.4.1
setuptools==58.0.4
spacy==3.8.3
textblob==0.19.0
torch==2.6.0
transformers==4.49.0
llama-cpp-python==0.3.8
nemoguardrails==0.13.0